+++ sumologic/resource_sumologic_cse_log_mapping.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"log"
+	"strconv"
 )
 
 func resourceSumologicCSELogMapping() *schema.Resource {
@@ -351,11 +352,17 @@ func resourceToCSELogMappingField(data interface{}) CSELogMappingField {
 		field.SkippedValues = resourceStringArrayToStringArray(fieldObj["skipped_values"].([]interface{}))
 		field.DefaultValue = fieldObj["default_value"].(string)
 		field.Format = fieldObj["format"].(string)
-		field.CaseInsensitive = fieldObj["case_insensitive"].(bool)
+		caseInsensitive := fieldObj["case_insensitive"]
+		if caseInsensitive != nil {
+			field.CaseInsensitive = fieldObj["case_insensitive"].(bool)
+		}
 		field.AlternateValues = resourceStringArrayToStringArray(fieldObj["alternate_values"].([]interface{}))
 		field.TimeZone = fieldObj["time_zone"].(string)
 		field.SplitDelimiter = fieldObj["split_delimiter"].(string)
-		field.SplitIndex = fmt.Sprint(fieldObj["split_index"].(int))
+		splitIndex := fieldObj["split_index"]
+		if splitIndex != nil {
+			field.SplitIndex = fmt.Sprint(splitIndex.(int))
+		}
 		field.FieldJoin = resourceStringArrayToStringArray(fieldObj["field_join"].([]interface{}))
 		field.JoinDelimiter = fieldObj["join_delimiter"].(string)
 		field.FormatParameters = resourceStringArrayToStringArray(fieldObj["format_parameters"].([]interface{}))
@@ -366,7 +373,7 @@ func resourceToCSELogMappingField(data interface{}) CSELogMappingField {
 			lookup = append(lookup, resourceToCSELogMappingLookUp([]interface{}{data}))
 		}
 
-		field.LookUp = lookup
+		field.LookUp = &lookup
 
 	}
 	return field
@@ -424,6 +431,7 @@ func setFields(d *schema.ResourceData, fields []CSELogMappingField) {
 	var f []map[string]interface{}
 
 	for _, t := range fields {
+
 		mapping := map[string]interface{}{
 			"name":              t.Name,
 			"value":             t.Value,
@@ -435,28 +443,35 @@ func setFields(d *schema.ResourceData, fields []CSELogMappingField) {
 			"alternate_values":  t.AlternateValues,
 			"time_zone":         t.TimeZone,
 			"split_delimiter":   t.SplitDelimiter,
-			"split_index":       t.SplitIndex,
 			"field_join":        t.FieldJoin,
 			"join_delimiter":    t.JoinDelimiter,
 			"format_parameters": t.FormatParameters,
 			"lookup":            getLookUpResource(t.LookUp),
 		}
+
+		splitIndex, err := strconv.Atoi(t.SplitIndex)
+		if err == nil {
+			mapping["split_index"] = splitIndex
+		}
+
 		f = append(f, mapping)
+
 	}
 
 	d.Set("fields", f)
 
 }
 
-func getLookUpResource(lookUp []CSELogMappingLookUp) []map[string]interface{} {
+func getLookUpResource(lookUp *[]CSELogMappingLookUp) []map[string]interface{} {
 	var s []map[string]interface{}
-
-	for _, l := range lookUp {
-		mapping := map[string]interface{}{
-			"key":   l.Key,
-			"value": l.Value,
+	if lookUp != nil {
+		for _, l := range *lookUp {
+			mapping := map[string]interface{}{
+				"key":   l.Key,
+				"value": l.Value,
+			}
+			s = append(s, mapping)
 		}
-		s = append(s, mapping)
 	}
 
 	return s