+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.31.2 (Unreleased)
 ENHANCEMENTS:
 * Add IsIncludedInDefaultSearch field to partition resource (GH-674)
+BUG FIXES:
+* Fix cse_log_mappings resource conversion affecting import  (GH-675)
 
 ## 2.31.1 (July 2, 2024)
 BUG FIXES: